# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TDI API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://tdi-api.test:81/api",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
        "/articleSsj2": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Shows article",
                "description": "Shows an article",
                "operationId": "showsArticles",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "Posts an article",
                "description": "Posts an article",
                "operationId": "postArticle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "title",
                                    "description",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "article title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "article description",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "article image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "article author",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/articleSsj2/{id}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Shows an article",
                "description": "Shows an article",
                "operationId": "showsArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "Updates an article",
                "description": "Updates an article",
                "operationId": "updateArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "title",
                                    "description",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "article title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "article description",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "article image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "article author",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "Deletes an article",
                "description": "Deletes an article",
                "operationId": "deleteArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Shows user",
                "description": "Shows an user",
                "operationId": "showsUsers",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Posts an user",
                "description": "Posts an user",
                "operationId": "postUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "title",
                                    "description",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "article title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "article description",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "article image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "article author",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Shows a user",
                "description": "Shows a user",
                "operationId": "showsUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Updates a user",
                "description": "Updates a user",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image",
                                    "title",
                                    "description",
                                    "user_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "article title",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "article description",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "article image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "user_id": {
                                        "description": "article author",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Deletes a user",
                "description": "Deletes a user",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/user/{id}/articles": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Shows a userArticles",
                "description": "Shows a userArticles",
                "operationId": "showsUserArticles",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    }
}